Create Key Signature Generator
Small data set of ordered accidentals as they appear on a musical staff. Program requires user to input() a valid starting pitch to generate the required amount of accidentals in that given key signature. Inputs that are not one of the 12 chromatic pitches available on a musical staff (or uncommon enharmonic such as "F-flat" or "E-sharp") will produce an error.